\name{mwindow-package}
\alias{mwindow-package}
\alias{mwindow}
\docType{package}
\title{
\packageTitle{mwindow}
}
\description{
\packageDescription{mwindow}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mwindow}
\packageIndices{mwindow}
This package is a simple collection of functions, based on lm(), to use linear interpolation
with a moving window.
This package contains the following functions:

moving_window()
check_parameter()
odd_wsize()
pvalue_control()

Details can be found in the man page of every function.
}
\author{
\packageAuthor{mwindow}

Maintainer: \packageMaintainer{mwindow}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
