\name{moving_window-package}
\alias{moving_window-package}
\alias{moving_window}
\docType{package}
\title{
\packageTitle{moving_window}
}
\description{
\packageDescription{moving_window}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{moving_window}
\packageIndices{moving_window}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{moving_window}

Maintainer: \packageMaintainer{moving_window}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
